Update local changes
+ Fix TextBlock size becoming too small
+ Move every segment of the code into separate parts
+ Fix MoveWriteFileInner() to write output data with incorrect size
+ Reduce refresh rate for the events to 100ms each
